Adding Coding Files
I added 4 coding file - the first is the Base Version which moves both the Horizontal Motor and the Vertical Motor, the second is a Multi-Photoresistor version which is more accurate when moving the motors, third is Version two which only moves the Horizontal Motor and then after it found the best angle only moves the Vertical Motor. Version 3 is a hybrid of the base version and Version 2 which moves the vertical motor a bit, then the horizontal motor a bit, and then repeats that process.